hint ints

+++ pdal/filters/PythonFilter.cpp
@@ -146,7 +146,8 @@ void PythonFilter::ready(PointTableRef table)
 PointViewSet PythonFilter::run(PointViewPtr view)
 {
     log()->get(LogLevel::Debug5) << "filters.python " << *m_script <<
-        " processing " << view->size() << " points." << std::endl;
+        " processing " << (int)view->size() << " points." << std::endl;
+
     m_pythonMethod->execute(view, getMetadata());
 
     PointViewSet viewSet;
